Become more type safe and less performant
Before this commit, isf.Int, isf.String and isf.Float64 were wrapper
types. This meant that type conversions happened when using these
as wrappers. For example, isf.String(123) was allowed. This is
convenient, but not type safe. Since goal of this module is to emulate
tagged unions, this was not a great solution.

This commit introduces overhead and boilerplate of making these symbols
functions and hiding wrapper types from user of this package. This
means that type conversions have to happen manually.

This results in more type safety, but more overhead and boilerplate.